removing UI folder
